refactor: paths and naming architecture
