+++ app/blocks/[block]/error.tsx
