+++ app/zkvm/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation"
 
-export default function Page() {
+export default function RedirectPage() {
   redirect("zkvms")
 }
