+++ components/Sidebar.tsx
@@ -71,7 +71,7 @@ export function AppSidebar() {
             </SidebarMenuItem>
             <SidebarMenuItem>
               <Link
-                href="/provers"
+                href="/teams"
                 className="inline-flex text-body hover:text-primary [&>svg:first-child]:me-2 [&>svg:first-child]:size-6"
               >
                 <ProofCheck />
