+++ components/sections/ProvingTeamsSection.tsx
@@ -10,7 +10,7 @@ import { getActiveClusters, getActiveMachineCount } from "@/lib/api/clusters"
 import { getClusterSummary, getTeamsSummary } from "@/lib/api/stats"
 import { transformClusters } from "@/lib/clusters"
 
-const ProversSection = async () => {
+const ProverTeamsSection = async () => {
   const teamsSummary = await getTeamsSummary()
   const clusterSummary = await getClusterSummary()
   const machineCount = await getActiveMachineCount()
@@ -53,7 +53,7 @@ const ProversSection = async () => {
       />
 
       <div className="flex justify-center">
-        <ButtonLink variant="outline" href="/provers">
+        <ButtonLink variant="outline" href="/teams">
           See all
         </ButtonLink>
       </div>
@@ -61,4 +61,4 @@ const ProversSection = async () => {
   )
 }
 
-export default ProversSection
+export default ProverTeamsSection
